+++ bot/settimezone.py
@@ -1,65 +1,66 @@
 #!/usr/bin/env python3
 """Conversation for deleting stored stickers."""
-from typing import cast, Dict
-from fuzzywuzzy import fuzz
+from typing import Dict, cast
+
 import pytz
 from telegram import (
-    Update,
-    Message,
-    InlineKeyboardMarkup,
+    Bot,
     InlineKeyboardButton,
+    InlineKeyboardMarkup,
     InlineQuery,
     InlineQueryResultArticle,
     InputTextMessageContent,
-    Bot,
+    Message,
+    Update,
     User,
 )
 from telegram.ext import (
-    ConversationHandler,
-    CommandHandler,
-    MessageHandler,
-    Filters,
-    InlineQueryHandler,
     ChosenInlineResultHandler,
+    CommandHandler,
+    ConversationHandler,
+    InlineQueryHandler,
+    MessageHandler,
+    filters,
 )
+from thefuzz import fuzz
 
 from bot.constants import REMOVE_KEYBOARD_KEY
 from bot.userdata import CCT, UserData
-from bot.utils import TIMEOUT_HANDLER, FALLBACK_HANDLER, remove_reply_markup
+from bot.utils import FALLBACK_HANDLER, TIMEOUT_HANDLER, remove_reply_markup
 
 STATE = 42
 
 
-def start(update: Update, context: CCT) -> int:
+async def start(update: Update, context: CCT) -> int:
     """Starts the conversation and asks for the timezone.
 
     Args:
         update: The Telegram update.
-        context: The callback context as provided by the dispatcher.
+        context: The callback context as provided by the application.
 
     Returns:
         int: The next state.
     """
     user_data = cast(UserData, context.user_data)
     message = cast(Message, update.effective_message)
-    message = message.reply_text(
+    message = await message.reply_text(
         f"Your current timezone is: {user_data.tzinfo}. Please press the button below to "
         "select a new timezone. You can scroll through the available options or type something to "
         "narrow down the options.",
         reply_markup=InlineKeyboardMarkup.from_button(
-            InlineKeyboardButton(text='Click me 👆', switch_inline_query_current_chat=''),
+            InlineKeyboardButton(text="Click me 👆", switch_inline_query_current_chat=""),
         ),
     )
     cast(Dict, context.chat_data)[REMOVE_KEYBOARD_KEY] = message
     return STATE
 
 
-def handle_inline_query(update: Update, _: CCT) -> int:
+async def handle_inline_query(update: Update, _: CCT) -> int:
     """Handles the inline query to show appropriate timezones.
 
     Args:
         update: The Telegram update.
-        _: The callback context as provided by the dispatcher.
+        _: The callback context as provided by the application.
 
     Returns:
         int: The next state.
@@ -73,7 +74,7 @@ def handle_inline_query(update: Update, _: CCT) -> int:
     else:
         timezones.sort()
 
-    inline_query.answer(
+    await inline_query.answer(
         results=[
             InlineQueryResultArticle(
                 id=tz, title=tz, input_message_content=InputTextMessageContent(tz)
@@ -87,12 +88,12 @@ def handle_inline_query(update: Update, _: CCT) -> int:
     return STATE
 
 
-def handle_timezone(update: Update, context: CCT) -> int:
+async def handle_timezone(update: Update, context: CCT) -> int:
     """Handles the selected timezone.
 
     Args:
         update: The Telegram update.
-        context: The callback context as provided by the dispatcher.
+        context: The callback context as provided by the application.
 
     Returns:
         int: The next state.
@@ -107,9 +108,9 @@ def handle_timezone(update: Update, context: CCT) -> int:
         message = cast(Message, update.effective_message)
         user_data.tzinfo = cast(str, message.text)
 
-    user.send_message(f'Your new timezone is {user_data.tzinfo}.')
+    await user.send_message(f"Your new timezone is {user_data.tzinfo}.")
 
-    remove_reply_markup(context)
+    await remove_reply_markup(context)
     return ConversationHandler.END
 
 
@@ -123,13 +124,13 @@ def build_set_timezone_conversation(bot: Bot) -> ConversationHandler:
         The ConversationHandler
     """
     return ConversationHandler(
-        entry_points=[CommandHandler('set_timezone', start)],
+        entry_points=[CommandHandler("set_timezone", start)],
         states={
             STATE: [
                 InlineQueryHandler(handle_inline_query),
                 # Use both CHRH and MH because we can't be sure which comes first …
                 ChosenInlineResultHandler(handle_timezone),
-                MessageHandler(Filters.via_bot(bot.id), handle_timezone),
+                MessageHandler(filters.ViaBot(bot.id), handle_timezone),
             ],
             ConversationHandler.TIMEOUT: [TIMEOUT_HANDLER],
         },