+++ bot/utils.py
@@ -1,14 +1,10 @@
 #!/usr/bin/env python3
 """Utility methods for the bot functionality."""
+from asyncio import Event
 from io import BytesIO
-from threading import Event
 from typing import cast
 
-from telegram import (
-    Update,
-    User,
-    Message,
-)
+from telegram import Message, Update, User
 from telegram.ext import ConversationHandler, TypeHandler
 
 from bot.constants import REMOVE_KEYBOARD_KEY
@@ -16,14 +12,16 @@ from bot.twitter import build_sticker
 from bot.userdata import CCT
 
 
-def get_sticker_photo_stream(text: str, user: User, context: CCT, event: Event = None) -> BytesIO:
+async def get_sticker_photo_stream(
+    text: str, user: User, context: CCT, event: Event = None
+) -> BytesIO:
     """
     Gives the sticker ID for the requested sticker.
 
     Args:
         text: The text to display on the tweet.
         user: The user the tweet is created for.
-        context: The callback context as provided by the dispatcher.
+        context: The callback context as provided by the application.
         event: Optional. If passed, ``event.is_set()`` will be checked before the time consuming
             parts of the sticker creation and if the event is set, the creation will be terminated.
 
@@ -31,53 +29,53 @@ def get_sticker_photo_stream(text: str, user: User, context: CCT, event: Event =
         Tuple[str, str]: The stickers unique file ID and file ID
     """
     sticker_stream = BytesIO()
-    sticker = build_sticker(text, user, context, event=event)
+    sticker = await build_sticker(text, user, context, event=event)
     sticker.save(sticker_stream, format="PNG")
     sticker_stream.seek(0)
 
     return sticker_stream
 
 
-def default_message(update: Update, _: CCT) -> None:
+async def default_message(update: Update, _: CCT) -> None:
     """
     Answers any message with a note that it could not be parsed.
 
     Args:
         update: The Telegram update.
-        _: The callback context as provided by the dispatcher.
+        _: The callback context as provided by the application.
     """
-    cast(Message, update.effective_message).reply_text(
+    await cast(Message, update.effective_message).reply_text(
         "Sorry, but I can only text messages. " 'Send "/help" for more information.'
     )
 
 
-def remove_reply_markup(context: CCT) -> None:
+async def remove_reply_markup(context: CCT) -> None:
     """Removes the reply markup of the message stored in ``context.chat_data[REMOVE_KEYBOARD]``,
     if any.
 
     Args:
-        context: The callback context as provided by the dispatcher.
+        context: The callback context as provided by the application.
     """
     if not context.chat_data:
         return
     message = context.chat_data.get(REMOVE_KEYBOARD_KEY, None)
     if isinstance(message, Message):
-        message.edit_reply_markup(None)
+        await message.edit_reply_markup(None)
 
 
-def conversation_timeout(update: Update, context: CCT) -> int:
+async def conversation_timeout(update: Update, context: CCT) -> int:
     """Informs the user that the operation has timed out, calls :meth:`remove_reply_markup` and
     ends the conversation.
 
     Args:
         update: The Telegram update.
-        context: The callback context as provided by the dispatcher.
+        context: The callback context as provided by the application.
 
     Returns:
         int: :attr:`telegram.ext.ConversationHandler.END`.
     """
-    cast(User, update.effective_user).send_message('Operation timed out. Aborting.')
-    remove_reply_markup(context)
+    await cast(User, update.effective_user).send_message("Operation timed out. Aborting.")
+    await remove_reply_markup(context)
 
     return ConversationHandler.END
 
@@ -87,19 +85,19 @@ TIMEOUT_HANDLER = TypeHandler(Update, conversation_timeout)
 conversations."""
 
 
-def conversation_fallback(update: Update, context: CCT) -> int:
+async def conversation_fallback(update: Update, context: CCT) -> int:
     """Informs the user that the input was invalid, calls :meth:`remove_reply_markup` and
     ends the conversation.
 
     Args:
         update: The Telegram update.
-        context: The callback context as provided by the dispatcher.
+        context: The callback context as provided by the application.
 
     Returns:
         int: The next state.
     """
-    cast(User, update.effective_user).send_message('Invalid input. Aborting operation.')
-    remove_reply_markup(context)
+    await cast(User, update.effective_user).send_message("Invalid input. Aborting operation.")
+    await remove_reply_markup(context)
 
     return ConversationHandler.END