+++ main.py
@@ -1,12 +1,13 @@
 #!/usr/bin/env python3
 """The script that runs the bot."""
+import asyncio
 import logging
 from configparser import ConfigParser
 
-from telegram import ParseMode
-from telegram.ext import Updater, Defaults, PicklePersistence, ContextTypes
+from telegram.constants import ParseMode
+from telegram.ext import Application, ContextTypes, Defaults, PersistenceInput, PicklePersistence
 
-from bot.setup import setup_dispatcher
+from bot.setup import setup_application
 from bot.userdata import UserData
 
 # Enable logging
@@ -15,7 +16,7 @@ logging.basicConfig(
     level=logging.INFO,
     filename="tsb.log",
 )
-aps_logger = logging.getLogger('apscheduler')
+aps_logger = logging.getLogger("apscheduler")
 aps_logger.setLevel(logging.WARNING)
 
 logger = logging.getLogger(__name__)
@@ -30,26 +31,30 @@ def main() -> None:
     admin_id = int(config["TwitterStatusBot"]["admins_chat_id"])
     sticker_chat_id = config["TwitterStatusBot"]["sticker_chat_id"]
 
-    defaults = Defaults(parse_mode=ParseMode.HTML, disable_notification=True, run_async=True)
+    defaults = Defaults(parse_mode=ParseMode.HTML, disable_notification=True, block=False)
     context_types = ContextTypes(user_data=UserData)
     persistence = PicklePersistence(
-        "tsb.pickle", context_types=context_types, single_file=False, store_chat_data=False
-    )
-    updater = Updater(
-        token,
-        defaults=defaults,
-        persistence=persistence,
-        workers=8,
+        "tsb.pickle",
         context_types=context_types,
+        single_file=False,
+        store_data=PersistenceInput(chat_data=False),
+    )
+    application = (
+        Application.builder()
+        .token(token)
+        .defaults(defaults)
+        .persistence(persistence)
+        .context_types(context_types)
+        .build()
     )
 
-    # Get the dispatcher to register handlers
-    dispatcher = updater.dispatcher
-    setup_dispatcher(dispatcher, admin_id, sticker_chat_id)
+    # Get the application to register handlers
+    asyncio.get_event_loop().run_until_complete(
+        setup_application(application, admin_id, sticker_chat_id)
+    )
 
     # Start the Bot
-    updater.start_polling(drop_pending_updates=True)
-    updater.idle()
+    application.run_polling(drop_pending_updates=True, close_loop=False)
 
 
 if __name__ == "__main__":